feat: Add merchant OCR mail
